Make row items not overlap participant lines.
Row items that may overlap participant lines are now given a modifier that
uses the Clear blendmode to clear the background behind the label composable.
This removes the participant lines beneath notes and labels with transparent
backgrounds so it doesn't look like the line runs through them. The entire
diagram is also then rendered in a compositing layer (using the alpha .99
trick since compositingStrategy isn't available in CfD yet), which ensures
that _only_ the participant lines are erased, and not whatever is drawn behind
the diagram itself.